+++ src/main/java/autolotto/machine/winning/WinningNumbers.java
@@ -1,4 +1,4 @@
-package autolotto;
+package autolotto.machine.winning;
 
 import java.util.ArrayList;
 import java.util.List;
