+++ src/test/java/autolotto/machine/winning/WinningTest.java
@@ -1,4 +1,4 @@
-package autolotto;
+package autolotto.machine.winning;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
